Update ID types from i32 to i16 in models

+++ src/models.rs
@@ -6,11 +6,11 @@ use diesel::prelude::*;
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct Customer {
     pub customer_id: i32,
-    pub store_id: i32,
+    pub store_id: i16,
     pub first_name: String,
     pub last_name: String,
     pub email: Option<String>,
-    pub address_id: i32,
+    pub address_id: i16,
     pub activebool: bool,
     pub create_date: NaiveDate,
     pub last_update: Option<chrono::NaiveDateTime>,
@@ -25,7 +25,7 @@ pub struct Address {
     pub address_col: String, // Corresponds to #[sql_name = "address"] in schema
     pub address2: Option<String>,
     pub district: String,
-    pub city_id: i32,
+    pub city_id: i16,
     pub postal_code: Option<String>,
     pub phone: String,
     pub last_update: chrono::NaiveDateTime,
